requirements_macos.txt
# Core dependencies - macOS compatible
numpy>=1.21.0
pillow>=8.3.0
opencv-python>=4.5.0
aiohttp>=3.8.0

# Mandatory OCR engines
paddleocr>=2.6.0
easyocr>=1.6.0

# PDF Processing  
PyMuPDF>=1.20.0

# Report Generation
markdown>=3.4.0

# Testing with pytest
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Structured logging for feedback
structlog>=22.3.0

# Google AI Integration (PRIMARY LLM provider)
google-genai>=0.1.0  # New unified Google GenAI SDK (Gemini API)

# Google Cloud AI (OPTIONAL - for Phase 3 image generation)
# google-cloud-aiplatform>=1.38.0  # Uncomment to enable Imagen image generation

# Note: vLLM has build issues on macOS - will use OpenAI API directly for now
# vllm>=0.2.0  # Uncomment for Linux/GPU systems